backend/requirements.txt
fastapi
uvicorn
pillow
pydentic
pydantic-settings
celery
python-multipart

# The following packages are installed in Dockerfile
# torch
# torchvision